+++ src/main/kotlin/icu/windea/pls/lang/actions/HandleUrlActionBase.kt
@@ -2,8 +2,11 @@ package icu.windea.pls.lang.actions
 
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.project.DumbAwareAction
+import icu.windea.pls.lang.selectGameType
 import icu.windea.pls.lang.tools.PlsUrlService
+import icu.windea.pls.model.ParadoxGameType
 import javax.swing.Icon
 
 abstract class HandleUrlActionBase(
@@ -30,6 +33,13 @@ abstract class HandleUrlActionBase(
     protected open fun isEnabled(e: AnActionEvent): Boolean = true
 
     protected abstract fun getTargetUrl(e: AnActionEvent): String?
+
+    protected fun getGameType(e: AnActionEvent): ParadoxGameType? {
+        val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
+        val gameType = selectGameType(file) ?: return null
+        if (gameType == ParadoxGameType.Core) return null
+        return gameType
+    }
 
     protected fun copyUrl(e: AnActionEvent) {
         val targetUrl = getTargetUrl(e) ?: return