+++ src/main/kotlin/icu/windea/pls/lang/actions/tools/CopyPathOrUrlActions.kt
@@ -6,7 +6,6 @@ import icu.windea.pls.core.orNull
 import icu.windea.pls.lang.actions.HandlePathActionBase
 import icu.windea.pls.lang.actions.HandleUrlActionBase
 import icu.windea.pls.lang.fileInfo
-import icu.windea.pls.lang.selectGameType
 import icu.windea.pls.lang.tools.PlsPathService
 import icu.windea.pls.lang.tools.PlsUrlService
 import icu.windea.pls.model.ParadoxRootInfo
@@ -26,8 +25,7 @@ interface CopyPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = copyPath(e)
 
         override fun getTargetPath(e: AnActionEvent): Path? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             return PlsPathService.getSteamGamePath(gameType.steamId, gameType.title)
         }
     }
@@ -36,8 +34,7 @@ interface CopyPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = copyPath(e)
 
         override fun getTargetPath(e: AnActionEvent): Path? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             return PlsPathService.getSteamWorkshopPath(gameType.steamId)
         }
     }
@@ -46,8 +43,7 @@ interface CopyPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = copyPath(e)
 
         override fun getTargetPath(e: AnActionEvent): Path? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             return PlsPathService.getGameDataPath(gameType.title)
         }
     }
@@ -90,8 +86,7 @@ interface CopyPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = copyUrl(e)
 
         override fun getTargetUrl(e: AnActionEvent): String? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             val steamId = gameType.steamId
             return PlsUrlService.getSteamGameStoreUrl(steamId)
         }
@@ -101,8 +96,7 @@ interface CopyPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = copyUrl(e)
 
         override fun getTargetUrl(e: AnActionEvent): String? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             val steamId = gameType.steamId
             return PlsUrlService.getSteamGameWorkshopUrl(steamId)
         }