+++ src/main/kotlin/icu/windea/pls/lang/actions/tools/OpenPathOrUrlActions.kt
@@ -6,7 +6,6 @@ import icu.windea.pls.core.orNull
 import icu.windea.pls.lang.actions.HandlePathActionBase
 import icu.windea.pls.lang.actions.HandleUrlActionBase
 import icu.windea.pls.lang.fileInfo
-import icu.windea.pls.lang.selectGameType
 import icu.windea.pls.lang.tools.PlsPathService
 import icu.windea.pls.lang.tools.PlsUrlService
 import icu.windea.pls.model.ParadoxRootInfo
@@ -26,8 +25,7 @@ interface OpenPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = openPath(e)
 
         override fun getTargetPath(e: AnActionEvent): Path? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             return PlsPathService.getSteamGamePath(gameType.steamId, gameType.title)
         }
     }
@@ -36,8 +34,7 @@ interface OpenPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = openPath(e)
 
         override fun getTargetPath(e: AnActionEvent): Path? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             return PlsPathService.getSteamWorkshopPath(gameType.steamId)
         }
     }
@@ -46,8 +43,7 @@ interface OpenPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = openPath(e)
 
         override fun getTargetPath(e: AnActionEvent): Path? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             return PlsPathService.getGameDataPath(gameType.title)
         }
     }
@@ -90,21 +86,9 @@ interface OpenPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = openUrl(e)
 
         override fun getTargetUrl(e: AnActionEvent): String? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             val steamId = gameType.steamId
             return PlsUrlService.getSteamGameStoreUrlInSteam(steamId)
-        }
-    }
-
-    class GameStorePageInSteamWebsite : HandleUrlActionBase() {
-        override fun actionPerformed(e: AnActionEvent) = openUrl(e)
-
-        override fun getTargetUrl(e: AnActionEvent): String? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
-            val steamId = gameType.steamId
-            return PlsUrlService.getSteamGameStoreUrl(steamId)
         }
     }
 
@@ -112,21 +96,9 @@ interface OpenPathOrUrlActions {
         override fun actionPerformed(e: AnActionEvent) = openUrl(e)
 
         override fun getTargetUrl(e: AnActionEvent): String? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
+            val gameType = getGameType(e) ?: return null
             val steamId = gameType.steamId
             return PlsUrlService.getSteamGameWorkshopUrlInSteam(steamId)
-        }
-    }
-
-    class GameWorkshopPageInSteamWebsite : HandleUrlActionBase() {
-        override fun actionPerformed(e: AnActionEvent) = openUrl(e)
-
-        override fun getTargetUrl(e: AnActionEvent): String? {
-            val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
-            val gameType = selectGameType(file) ?: return null
-            val steamId = gameType.steamId
-            return PlsUrlService.getSteamGameWorkshopUrl(steamId)
         }
     }
 
@@ -148,7 +120,27 @@ interface OpenPathOrUrlActions {
         }
     }
 
-    class ModPageInSteamWebsite : HandleUrlActionBase() {
+    class GameStorePage : HandleUrlActionBase() {
+        override fun actionPerformed(e: AnActionEvent) = openUrl(e)
+
+        override fun getTargetUrl(e: AnActionEvent): String? {
+            val gameType = getGameType(e) ?: return null
+            val steamId = gameType.steamId
+            return PlsUrlService.getSteamGameStoreUrl(steamId)
+        }
+    }
+
+    class GameWorkshopPage : HandleUrlActionBase() {
+        override fun actionPerformed(e: AnActionEvent) = openUrl(e)
+
+        override fun getTargetUrl(e: AnActionEvent): String? {
+            val gameType = getGameType(e) ?: return null
+            val steamId = gameType.steamId
+            return PlsUrlService.getSteamGameWorkshopUrl(steamId)
+        }
+    }
+
+    class ModPage : HandleUrlActionBase() {
         override fun isVisible(e: AnActionEvent): Boolean {
             val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return false
             val fileInfo = file.fileInfo ?: return false
@@ -166,3 +158,4 @@ interface OpenPathOrUrlActions {
         }
     }
 }
+