+++ src/main/kotlin/icu/windea/pls/lang/tools/PlsGitService.kt
@@ -1,52 +1,29 @@
 package icu.windea.pls.lang.tools
 
 import icu.windea.pls.core.console.CommandExecutionException
-import icu.windea.pls.core.executeCommand
-import icu.windea.pls.core.toPath
+import icu.windea.pls.lang.tools.impl.PlsGitServiceImpl
 import java.io.IOException
 import java.nio.file.Path
-import kotlin.io.path.exists
 
-object PlsGitService {
-    fun getRepositoryPathFromUrl(url: String): String {
-        return url.substringAfterLast('/').removeSuffix(".git")
-    }
+interface PlsGitService {
+    fun getRepositoryPathFromUrl(url: String): String
 
-    fun isUpdateToDate(message: String?): Boolean {
-        return message.isNullOrEmpty() || message.startsWith("Already up to date")
-    }
+    fun isUpdateToDate(message: String?): Boolean
 
     @Throws(IOException::class, InterruptedException::class, CommandExecutionException::class)
-    fun checkRemote(url: String): String {
-        return lsRemote(url)
-    }
+    fun checkRemote(url: String): String
 
     @Throws(IOException::class, InterruptedException::class, CommandExecutionException::class)
-    fun syncFromRemote(url: String, parentDirectory: String): String {
-        val localRepoDirectory = getRepositoryPathFromUrl(url)
-        val parentPath = parentDirectory.toPath()
-        val localRepoPath = parentPath.resolve(localRepoDirectory)
-        if (!localRepoPath.exists()) {
-            return clone(url, parentPath)
-        } else {
-            return pull(url, localRepoPath)
-        }
-    }
+    fun syncFromRemote(url: String, parentDirectory: String): String
 
     @Throws(IOException::class, InterruptedException::class, CommandExecutionException::class)
-    fun lsRemote(url: String): String {
-        return executeCommand("git ls-remote $url")
-    }
+    fun lsRemote(url: String): String
 
     @Throws(IOException::class, InterruptedException::class, CommandExecutionException::class)
-    fun clone(url: String, workDirectory: Path): String {
-        val wd = workDirectory.normalize().toAbsolutePath().toFile()
-        return executeCommand("git clone $url", workDirectory = wd)
-    }
+    fun clone(url: String, workDirectory: Path): String
 
     @Throws(IOException::class, InterruptedException::class, CommandExecutionException::class)
-    fun pull(url: String, workDirectory: Path): String {
-        val wd = workDirectory.normalize().toAbsolutePath().toFile()
-        return executeCommand("git pull $url", workDirectory = wd)
-    }
+    fun pull(url: String, workDirectory: Path): String
+
+    companion object : PlsGitService by PlsGitServiceImpl()
 }